requirements.txt
asgiref==3.8.1
black==24.4.0
click==8.1.7
Django==5.0.4
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
djhtml==3.0.6
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
python-dateutil==2.9.0.post0
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.5.0
